+++ MaxMind.MinFraud.UnitTest/Request/EmailTest.cs
@@ -17,6 +17,14 @@ namespace MaxMind.MinFraud.UnitTest.Request
         }
 
         [Fact]
+        public void TestAddressMD5WithUppercase()
+        {
+            var email = new Email("Test@maxmind.com");
+            Assert.Equal("977577b140bfb7c516e4746204fbdb01", email.AddressMD5);
+        }
+
+
+        [Fact]
         public void TestInvalidAddress()
         {
             Assert.Throws<ArgumentException>(() => new Email(address: "no-domain"));